refactor sat package into more files and use Atmo's error handling behaviour for standalone HTTP server mode
